requirements.txt
Django>=4.2,<6.0
Markdown>=3.4
python-dotenv>=1.0
gunicorn>=21.0
